chore: remove duplicate definition api_auth
