+++ dozer-api/src/lib.rs
@@ -1,8 +1,7 @@
-mod api_auth;
 pub mod api_generator;
 mod api_helper;
 pub mod api_server;
-pub mod auth;
+mod auth;
 mod errors;
 pub mod generator;
 mod grpc;
