+++ dozer-api/src/tests/routes.rs
@@ -1,6 +1,6 @@
 use super::super::api_server::ApiServer;
 use crate::{
-    api_auth::ApiSecurity, api_server::CorsOptions, generator::oapi::generator::OpenApiGenerator,
+    api_server::CorsOptions, auth::api::ApiSecurity, generator::oapi::generator::OpenApiGenerator,
     test_utils,
 };
 use dozer_types::serde_json::{json, Value};
